Added profile picture service and removed put profile picture from user service
