+++ src/app/services/user.service.ts
@@ -45,13 +45,6 @@ export class UserService {
     return this.putUserRequest(userId, token, updateUserFormGroup, languages, technologies, userRoles, userFriends);
   }
 
-  putProfilePictureFromSessionStorageRequest(newPicture: File): Observable<object> {
-    const userId = this._tokenService.getUserIdFromSessionStorageToken();
-    const token = this._tokenService.getTokenFromSessionStorage();
-
-    return this.putProfilePictureRequest(userId, token, newPicture);
-  }
-
   putBareUserFromSessionStorageRequest(user: User, password: string): Observable<object> {
     const userId = this._tokenService.getUserIdFromSessionStorageToken();
     const token = this._tokenService.getTokenFromSessionStorage();
@@ -148,16 +141,6 @@ export class UserService {
       headers: new HttpHeaders().set('Authorization', 'Bearer ' + authToken)
     };
     return this._http.put(AppConstants.API_USER_URL, body, options);
-  }
-
-  putProfilePictureRequest(userId: Guid, authToken: string, newPicture: File): Observable<object> {
-    const form = new FormData();
-    form.append('picture', newPicture);
-    const options = {
-      params: new HttpParams().set('UserId', userId.toString()),
-      headers: new HttpHeaders().set('Authorization', 'Bearer ' + authToken)
-    };
-    return this._http.put(AppConstants.API_USER_URL + '/ProfilePicture', form, options);
   }
 
   deleteUserRequest(userId: Guid, authToken: string): Observable<object> {
